+++ CacheHashTable/CacheLRU.cpp
@@ -25,9 +25,9 @@ void CacheLRU<HashT, MapT>::display_trackers(double time)
     std::cout << CAPACITY << ",";                           // model parameter
     std::cout << (double) time/t_search << ",";             // latency (ns)
     std::cout << (double) t_hit/t_search*100 << ",";        // hitrate (%)
-    std::cout << (double) size() << ",";                    // # pairs
-    std::cout << (double) content_size() << ",";            // content size (B)
-    std::cout << (double) bookkeeping_overhead();           // overhead (B)
+    std::cout << size() << ",";                             // # pairs
+    std::cout << content_size() << ",";                     // content size (B)
+    std::cout << bookkeeping_overhead();                    // overhead (B)
     std::cout << std::endl;
 }
